Move FailState doctests to unit tests
